add service for logging, only tokio-std for now
